Fix analysis findings and prepare local Sonar runs

+++ ASFWDriver/Isoch/Audio/AudioIOPath.cpp
@@ -24,8 +24,9 @@ void MaybeDrainRxStartup(AudioIOPathState& state) {
     *state.rxStartupDrained = true;
 }
 
+// NOLINTNEXTLINE(bugprone-easily-swappable-parameters)
 kern_return_t HandleBeginRead(AudioIOPathState& state,
-                              uint32_t ioBufferFrameSize,
+                              uint32_t ioBufferFrameSize, // NOLINT(bugprone-easily-swappable-parameters)
                               uint64_t sampleTime) {
     if (!state.inputBuffer) {
         return kIOReturnNotReady;
@@ -58,20 +59,20 @@ kern_return_t HandleBeginRead(AudioIOPathState& state,
     MaybeDrainRxStartup(state);
 
     if (state.rxQueueValid && state.rxQueueReader) {
-        auto* pcmFirst = reinterpret_cast<int32_t*>(segment.address + offsetBytes);
-        const uint32_t read1 = state.rxQueueReader->Read(pcmFirst, firstFrames);
-        if (read1 < firstFrames) {
-            memset(pcmFirst + read1 * ch,
-                   0,
-                   size_t(firstFrames - read1) * sizeof(int32_t) * ch);
-        }
+            auto* pcmFirst = reinterpret_cast<int32_t*>(segment.address + offsetBytes);
+            const uint32_t read1 = state.rxQueueReader->Read(pcmFirst, firstFrames);
+            if (read1 < firstFrames) {
+                memset(pcmFirst + (static_cast<size_t>(read1) * ch),
+                       0,
+                       size_t(firstFrames - read1) * sizeof(int32_t) * ch);
+            }
 
         if (secondFrames > 0) {
             auto* pcmSecond = reinterpret_cast<int32_t*>(segment.address);
             if (read1 == firstFrames) {
                 const uint32_t read2 = state.rxQueueReader->Read(pcmSecond, secondFrames);
                 if (read2 < secondFrames) {
-                    memset(pcmSecond + read2 * ch,
+                    memset(pcmSecond + (static_cast<size_t>(read2) * ch),
                            0,
                            size_t(secondFrames - read2) * sizeof(int32_t) * ch);
                 }
@@ -230,8 +231,9 @@ kern_return_t HandleWriteEnd(AudioIOPathState& state,
 
 } // namespace detail
 
+// NOLINTNEXTLINE(bugprone-easily-swappable-parameters)
 kern_return_t HandleIOOperation(AudioIOPathState& state,
-                                IOUserAudioIOOperation operation,
+                                IOUserAudioIOOperation operation, // NOLINT(bugprone-easily-swappable-parameters)
                                 uint32_t ioBufferFrameSize,
                                 uint64_t sampleTime) {
     switch (operation) {
